1.0 + 1.0.1 update
Рефакторинг кода, Исправлены баги и прочее

+++ app/src/main/java/com/sinya/projects/sportsdiary/data/database/entity/DataProportions.kt
@@ -30,5 +30,5 @@ import androidx.room.Index
 data class DataProportions(
     @ColumnInfo(name = "proportion_id") val proportionId: Int,
     @ColumnInfo(name = "type_id") val typeId: Int,
-    @ColumnInfo(name = "value", defaultValue = "0") val value: Int = 0
+    @ColumnInfo(name = "value", defaultValue = "0") val value: Float = 0f
 )
